build fix
svn:sync-xref-src-repo-rev: r73716

+++ src/VBox/Runtime/common/dbg/dbgmoddwarf.cpp
@@ -1,4 +1,4 @@
-/* $Id: dbgmoddwarf.cpp 38557 2011-08-29 13:13:27Z knut.osmundsen@oracle.com $ */
+/* $Id: dbgmoddwarf.cpp 38558 2011-08-29 13:19:17Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - Debug Info Reader For DWARF.
  */
@@ -822,7 +822,7 @@ typedef struct RTDWARFLINESTATE
     struct
     {
         uint64_t        uAddress;
-        uint32_t        idxOp;
+        uint64_t        idxOp;
         uint32_t        iFile;
         uint32_t        uLine;
         uint32_t        uColumn;
@@ -873,7 +873,7 @@ static int rtDwarfLine_DefineFileName(PRTDWARFLINESTATE pLnState, const char *ps
     /*
      * Resize the array if necessary.
      */
-    size_t iFileName = pLnState->cFileNames;
+    uint32_t iFileName = pLnState->cFileNames;
     if ((iFileName % 2) == 0)
     {
         void *pv = RTMemRealloc(pLnState->papszFileNames, sizeof(pLnState->papszFileNames[0]) * (iFileName + 2));
